02-starting-openai-lib/requirements.txt
ipykernel
openai>=1.0.0
python-dotenv
yfinance
pandas
matplotlib
yfinance
requests
Pillow
